% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.datatype.R
\name{check.datatype}
\alias{check.datatype}
\title{If it is an integer, change it to numeric and check the character and numeric variables
After roading the data, you must name the data as 'data'!!!!}
\usage{
check.datatype()
}
\arguments{
\item{data}{input data}
}
\value{
You can check the type for each variable.
}
\description{
If it is an integer, change it to numeric and check the character and numeric variables
After roading the data, you must name the data as 'data'!!!!
}
\examples{
check.datatype()
enter : data
enter : colnumber

}
