% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_test.R
\name{numeric_test}
\alias{numeric_test}
\title{Check variable type}
\usage{
numeric_test(data)
}
\arguments{
\item{data}{input data}
}
\value{
You can check the type for each variable.
}
\description{
If it is an integer, change it to numeric and check the character and numeric variables
}
\examples{
numeric_test(datasetname)
}
