\name{Whale-package}
\alias{Whale-package}
\alias{Whale}
\docType{package}
\title{
\packageTitle{Whale}
}
\description{
\packageDescription{Whale}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Whale}
\packageIndices{Whale}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{Whale}

Maintainer: \packageMaintainer{Whale}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
