% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohwhaley.R, R/whale.R
\name{say}
\alias{say}
\title{Summon a Whale for a customisable pick-me-up}
\usage{
say(what)

say(what)
}
\arguments{
\item{what}{Whatever phrase you want to echo. If not supplied, a random phrase is chosen.}
}
\value{
Whale shaped message

Whale shaped message
}
\description{
Summon a Whale for a customisable pick-me-up

Summon a Whale for a customisable pick-me-up
}
\examples{
say("hi")
say("hi")
}
